refactor(dashboard): extract RecommendationModal into colocated component
First step decomposing the 2439-line dashboard god component. Moves the
self-contained RecommendationModal (no closure over page state) verbatim
into _components/recommendation-modal.tsx and removes the now-unused
modal-only imports (Dialog, TrailerEmbed, MusicPreview, WhyThisPick).
Behavior-identical; page.tsx 2439->2328 LOC. Establishes the _components
colocation pattern for subsequent steps.
